resources/cdk/photo_asset_manager/frontend/requirements.txt
aws-cdk-lib==2.67.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.79
types-boto3
